restructure, fix stylesheet file name so it renders
