+++ app.js
@@ -1,0 +1,64 @@
+var express = require('express'),
+    path = require('path'),
+    favicon = require('serve-favicon'),
+    logger = require('morgan'),
+    cookieParser = require('cookie-parser'),
+    bodyParser = require('body-parser');
+
+var routes = require('./app/routes/index');
+var users = require('./app/routes/users');
+
+var app = express();
+
+app.set('views', path.join(__dirname, 'app/views'));
+app.set('view engine', 'jade');
+
+// uncomment after placing your favicon in /public
+//app.use(favicon(__dirname + '/public/favicon.ico'));
+app.use(logger('dev'));
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(cookieParser());
+app.use(express.static(path.join(__dirname, 'public')));
+
+app.get('/', function(req, res, next) {
+  console.log('routing to index');
+  res.render('index');
+});
+
+// catch 404 and forward to error handler
+// This is due to the way express attemps to match a requested
+// path: since no other paths can match at this point, the app
+// will default to handling this function.
+app.use(function(req, res, next) {
+  console.log('routing to NOT index');
+var err = new Error('Not Found');
+err.status = 404;
+  next(err);
+});
+
+// error handlers
+
+// development error handler
+// will print stacktrace
+if (app.get('env') = == 'development') {
+  app.use(function(err, req, res, next) {
+    res.status(err.status || 500);
+    res.render('error', {
+      message: err.message,
+      error: err
+    });
+  });
+}
+
+// production error handler
+// no stacktraces leaked to user
+app.use(function(err, req, res, next) {
+  res.status(err.status || 500);
+  res.render('error', {
+    message: err.message,
+    error: {}
+  });
+});
+
+module.exports = app;
