+++ app/views/layout.jade
@@ -1,6 +1,6 @@
 doctype html
 html
   head
-    link(rel='stylesheet', href='/styles/style.css')
+    link(rel='stylesheet', href='/styles/site.css')
   body
     block content
